feat: Implement palindrome partitioning algorithm
This commit implements the palindrome partitioning algorithm in the `partition` method of the `Solution` class. The algorithm uses recursion to find all possible palindrome partitions of a given string `s`. It checks if each substring is a palindrome and adds it to the result list if it is. The algorithm has been tested and is working correctly.